fix(Transaction & Asset Type): Change from uint8 to uint16

+++ contracts/Governance.sol
@@ -42,12 +42,12 @@ import "./interfaces/IGovernance.sol";
 contract Governance is IGovernance, Ownable {
   event SetContractGuard(address extContract, address guardAddress);
 
-  event SetAssetGuard(uint8 assetType, address guardAddress);
+  event SetAssetGuard(uint16 assetType, address guardAddress);
 
   // Transaction Guards
 
   mapping(address => address) public override contractGuards;
-  mapping(uint8 => address) public override assetGuards;
+  mapping(uint16 => address) public override assetGuards;
 
   // Transaction Guards
 
@@ -64,11 +64,11 @@ contract Governance is IGovernance, Ownable {
     emit SetContractGuard(extContract, guardAddress);
   }
 
-  function setAssetGuard(uint8 assetType, address guardAddress) external onlyOwner {
+  function setAssetGuard(uint16 assetType, address guardAddress) external onlyOwner {
     _setAssetGuard(assetType, guardAddress);
   }
 
-  function _setAssetGuard(uint8 assetType, address guardAddress) internal {
+  function _setAssetGuard(uint16 assetType, address guardAddress) internal {
     require(guardAddress != address(0), "Invalid guardAddress");
 
     assetGuards[assetType] = guardAddress;
