+++ contracts/PoolFactory.sol
@@ -167,26 +167,24 @@ contract PoolFactory is
   ) external returns (address) {
     require(_supportedAssets.length <= _maximumSupportedAssetCount, "maximum assets reached");
 
-    bytes memory poolLogicData =
-      abi.encodeWithSignature(
-        "initialize(address,bool,string,string)",
-        address(this),
-        _privatePool,
-        _fundName,
-        _fundSymbol
-      );
+    bytes memory poolLogicData = abi.encodeWithSignature(
+      "initialize(address,bool,string,string)",
+      address(this),
+      _privatePool,
+      _fundName,
+      _fundSymbol
+    );
 
     address fund = deploy(poolLogicData, 2);
 
-    bytes memory managerLogicData =
-      abi.encodeWithSignature(
-        "initialize(address,address,string,address,(address,bool)[])",
-        address(this),
-        _manager,
-        _managerName,
-        fund,
-        _supportedAssets
-      );
+    bytes memory managerLogicData = abi.encodeWithSignature(
+      "initialize(address,address,string,address,(address,bool)[])",
+      address(this),
+      _manager,
+      _managerName,
+      fund,
+      _supportedAssets
+    );
 
     address managerLogic = deploy(managerLogicData, 1);
     // Ignore return value as want it to continue regardless
@@ -365,7 +363,7 @@ contract PoolFactory is
     return IAssetHandler(_assetHandler).getUSDPrice(asset);
   }
 
-  function getAssetType(address asset) external view override returns (uint8) {
+  function getAssetType(address asset) external view override returns (uint16) {
     return IAssetHandler(_assetHandler).assetTypes(asset);
   }
 
@@ -420,12 +418,12 @@ contract PoolFactory is
     assembly {
       let succeeded := delegatecall(gas(), pool, add(_data, 0x20), mload(_data), 0, 0)
       switch iszero(succeeded)
-        case 1 {
-          // throw if delegatecall failed
-          let size := returndatasize()
-          returndatacopy(0x00, 0x00, size)
-          revert(0x00, size)
-        }
+      case 1 {
+        // throw if delegatecall failed
+        let size := returndatasize()
+        returndatacopy(0x00, 0x00, size)
+        revert(0x00, size)
+      }
     }
     emit LogUpgrade(msg.sender, pool);
 
@@ -466,7 +464,7 @@ contract PoolFactory is
 
   function getGuard(address extContract) external view override returns (address) {
     if (isValidAsset(extContract)) {
-      uint8 assetType = IAssetHandler(_assetHandler).assetTypes(extContract);
+      uint16 assetType = IAssetHandler(_assetHandler).assetTypes(extContract);
       return IGovernance(governanceAddress).assetGuards(assetType);
     }
     return IGovernance(governanceAddress).contractGuards(extContract);
