+++ contracts/interfaces/IHasAssetInfo.sol
@@ -39,7 +39,7 @@ interface IHasAssetInfo {
 
   function getAssetPrice(address asset) external view returns (uint256);
 
-  function getAssetType(address asset) external view returns (uint8);
+  function getAssetType(address asset) external view returns (uint16);
 
   function getMaximumSupportedAssetCount() external view returns (uint256);
 }
